requirements.txt
langchain
chromadb
openai
python-dotenv
tiktoken
fastapi